requirements.txt
selenium
pytesseract
Pillow
requests